requirements.txt
# Core LLM — Google Gemini
google-genai>=1.0.0

# Agent Orchestration
langgraph>=0.1.0

# RAG — Text Splitting
langchain-text-splitters>=0.2.0

# Vector Store
chromadb>=0.5.0

# Memory / Database
sqlalchemy>=2.0.0

# UI
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
tenacity>=8.0.0
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
